Rework parsed material inherintance completely (#4549)
