+++ modules/phase_field/src/materials/DerivativeSumMaterial.C
@@ -16,6 +16,7 @@ InputParameters validParams<DerivativeSumMaterial>()
 
 DerivativeSumMaterial::DerivativeSumMaterial(const std::string & name,
                                              InputParameters parameters) :
+    FunctionMaterialBase(name, parameters),
     DerivativeBaseMaterial(name, parameters),
     _sum_materials(getParam<std::vector<std::string> >("sum_materials")),
     _num_materials(_sum_materials.size())
