+++ modules/phase_field/src/materials/MathFreeEnergy.C
@@ -11,10 +11,12 @@ InputParameters validParams<MathFreeEnergy>()
 
 MathFreeEnergy::MathFreeEnergy(const std::string & name,
                        InputParameters parameters) :
+    FunctionMaterialBase(name, parameters),
     DerivativeBaseMaterial(name, parameters),
     _c(coupledValue("c")),
     _c_var(coupled("c"))
-{}
+{
+}
 
 Real
 MathFreeEnergy::computeF()
